Move all form variations into one form component
